implemented a crude acoustic target cost. This slows down the synthesis because currently, the unit F0 is computed online by accessing the respective unit audio data. This should be done offline, as part of the feature vectors.
But the result sounds nice, more regular than the purely linguistic target costs.
(of course, only when using an F0 and duration prediction which fits the voice!)

Acoustic target costs are currently disabled by default; to turn them on, use config setting

voice.xyz.acoustictargetcost.weight = 0.5

With a value > 0 (but <= 1).




git-svn-id: https://mary.opendfki.de/repos/trunk@544 953a6561-930b-0410-b2a6-db37d1b2ae63