solution to ticket:133 -- created a new datatype REALISED_ACOUSTPARAMS which is produced from AUDIO data -- the trick is to store the XML tree in the AUDIO mary data object, together with the actual audio data.
Targets are adapted to store a reference to a Maryxml Element in addition to an utterance Item;
the way these are filled at the moment is that Items in the Segment relation store a reference
to the DOM Element they were created from, so the createTargets() code has an easy job in trying
to align them. This means that from now on, we can start implementing XML-based target feature
processors, in order to gradually remove the dependency of unit selection code on Utterance
structures.

UnitSelectionSynthesizer, after creating audio, queries the concatenation data objects filled
by the unit concatenator, and updates phone and boundary durations in the XML file accordingly.

Tested with unit selection only; should work with other WaveformSynthesizers as well, just
for them, the same values as for ACOUSTPARAMS will be output in REALISED_ACOUSTPARAMS.

The current method means that when requesting REALISED_ACOUSTPARAMS in addition to AUDIO, 
all processing steps will need to be computed twice; in the future, a caching mechanism can
be used to make this more efficient.



git-svn-id: https://mary.opendfki.de/repos/trunk@827 953a6561-930b-0410-b2a6-db37d1b2ae63