alternative implementation to _process(), named processTargetFeatures(), and not used by default yet.
processTargetFeatures() uses the FeatureDefinition and FeatureVector classes. It is modular
in the sense that (different from _process()) each feature vector line is processed separately,
in a public method "features2context()" which can also be called externally.
This method relies on the presence of the features "mary_prev_prev_phoneme" and 
"mary_next_next_phoneme" to generate proper quintphones; if they are missing, only triphones
are generated, and "0" is used as the next-next and prev-prev phoneme symbol.

git-svn-id: https://mary.opendfki.de/repos/trunk@1089 953a6561-930b-0410-b2a6-db37d1b2ae63