Thorough reworking of modules and datatypes, ticket:174.
1. MaryData has a type (which is now language-independent) and a locale.

2. Modules know about their locale, which can be null, a language (e.g., "en"), or language and country (e.g., "en_US").

3. Modules are administered in a new class, marytts.modules.ModuleRegistry. Its main function is to
find the sequence of modules needed to convert a certain input type (given a locale and a voice) into an output type. To do that, it maintains a multi-key map, associating modules with input type, locale (optional), and voice (optional) for which it is suitable. When looking up modules in order to find a path from input to output type, preference is given to more specific modules (e.g., with the right locale and voice) over more generic (e.g., language-independent) modules. This allows us to provide baseline modules in a language-independent manner where that is appropriate, and still it is easy to override these with language-specific modules. 

In particular, if there is data from a locale describing language and country (e.g., "en_US"), the registry will look for country-specific modules and, if none are found, will use language-specific ones. This allows us to use certain components (e.g., the tokenizer) for all dialects of English, but have country-specific phonemiser and allophone generator modules. This reduces the load to add new variants to an existing language.

This supersedes the ad hoc lookup which was formerly done in the Mary class itself. It will also supersede the preferred modules list in the Voice class -- any modules which should be used specifically for a given voice can be handled by adding a voice-specific record in the ModuleRegistry.

4. MaryDataType was changed significantly.

- MaryDataType has no locale information anymore. Therefore, a distinction between TEXT_DE, TEXT_EN etc. makes no sense anymore.

- There is a centrally held list of generic data types, which should be used by all languages if it makes sense to use them. The core set is:
   TEXT, RAWMARYXML, TOKENS, WORDS, PARTSOFSPEECH, PHONEMES, INTONATION, ALLOPHONES, ACOUSTPARAMS, and AUDIO, and REALISED_ACOUSTPARAMS.
There are some more data types that may be used in certain circumstances, such as various alternative input formats (SSML, SABLE, APML, SIMPLEPHONEMES), and various intermediate output formats (TARGETFEATURES, FESTIVAL_UTT, MBROLA, etc.).

- The method for defining and for accessing MaryDataType objects has also changed. 

  -> Definition: the indirect definer classes (e.g., TEXT_Definer) are not used anymore. Instead, every installation component (marybase, german, english, tibetan, any-new-language) defines its own "constant objects" (public static final MaryDataType) using a simple MaryDataType constructor.

  -> Access: Instead of the indirect access via MaryDataType.get("TEXT"), modules must access the constant objects directly: MaryDataType.TEXT, USEnglishDataTypes.FREETTS_TOKENS, etc. This helps find errors at compile time.

- Example texts are not part of the mary data type anymore, because the examples depend on a locale but mary data types are locale-independent. Instead, examples are held as text resources, following the following naming convention:

  -> for locale-independent examples (locale == null): <package of the type>.NAME.example (e.g., for marytts.datatypes.MaryDataType.MBROLA, the example will be looked for at marytts.datatypes.MBROLA.example).
  -> for locale-specific examples: marytts.language.<locale>.datatypes.NAME.<locale>.example (e.g., for a US-English example for the type TEXT, the code will look for marytts.language.en_US.datatypes.TEXT.en_US.example).
  
Whether an example is available for a given type and locale is thus determined at runtime -- if an example can be found in the classpath at the location described above, it is used; otherwise there is no example.

5. There are minor changes in the client-server protocol.

- Requests from client to server: the client must specify the LOCALE=<locale> in the request, after IN= and OUT=. The current GUI client determines the locale from the voice used.

- the list of data types does not contain locale information anymore.

- requests for example texts must include a locale, as in MARY EXAMPLETEXT SSML en_US.


6. As a result of these changes, there is a strict distinction between locale "en" and "en_US" now; the old data files using "en" will not work anymore, except for conversions involving only modules with locale ENGLISH rather than US (e.g, you can convert a general xml:lang="en" document from RAWMARYXML to PARTSOFSPEECH, but not to PHONEMES). Simple solution: replace with xml:lang="en-US". (not the dash: in xml:lang, language and country must be separated by a dash, in package names, it must be an underscore).



The new setup seems to be basically working, but has only been briefly tested. I am sure there are various devils in various details, but we will sort them out over time.

git-svn-id: https://mary.opendfki.de/repos/trunk@1361 953a6561-930b-0410-b2a6-db37d1b2ae63