First fully functional version of client/server communication with client requests. I kept the older versions of client and server and put new http classes under corresponding http packages, i.e. marytts.server.http, marytts.client.http. Therefore, the old version that uses socket based communication is still fully functional. You can run the new versions by simply changing the main classes in run targets while executing server and client as follows: Main class for server: marytts.server.Mary --> marytts.server.http.MaryHttp Main class for GUI client: marytts.client.MaryGUIClient --> marytts.client.http.MaryGUIHttpClient
All major features work OK using HTTP requests:
- sending/receiving info from/to server
- receiving and playing/saving audio from server
Part of the error and warning log mechanisms might be broken since it was not possible to directly include them in HTTP communication. The code is still there as commented out in order not to forget adding similar functionality.
This does not apply to old version of client and server, of course - they run as before.

git-svn-id: https://mary.opendfki.de/repos/trunk@1507 953a6561-930b-0410-b2a6-db37d1b2ae63