Major cleanup in the prediction of target features, ticket #181. THIS BREAKS COMPATIBILITY, PREVIOUS VOICES WILL NOT WORK ANYMORE!
1. The major change itself

Concretely, I have changed all feature processors in marytts.features.MaryGenericFeatureProcessors and marytts.features.MaryLanguageFeatureProcessors so that they get the information directly from the MARYXML structure. There is no conversion to Utterance/Relation/Item format done anymore at the level of ACOUSTPARAMS, but the TargetFeatureLister and Synthesis modules access directly ACOUSTPARAMS input.

Consequently, XML2UttAcoustParams modules can be removed from the runtime, and ultimately from the source as well.

Also, the phone set used for computing phone features is now the same as the one used in the MARYXML "ph" attributes (e.g., SAMPA for German and English). No conversion to a freetts phoneset anymore at this level.

Finally, I have also changed the names of all features, by removing the "mary_" prefix, which several years ago was intended to be a temporary way of distinguishing these from the FREETTS feature processors. As a result, in all existing Feature definition and tree files, we must also remove the mary_ prefix from feature names. This is done in this commit for the "fallback" trees for CARTDurationModeller and CARTF0Modeller, which are lying in lib/modules/en/cap. References to feature names throughout the java code have also been verified and corrected.

2. A JUnit 4 example test file

We should use JUnit 4 for writing "unit tests", an automated way of verifying that the code fulfils certain assertions. I have written a test file marytts.test.junit4.FeatureProcessorTest, which does a full verification of all TargetElementNavigators and, for the moment, just one specific FeatureProcessor. The JUnit tests can be run directly in Eclipse, by defining a "JUnit test" run target. This particular test needs a VM argument "-Dmary.base=/path/to/mary/runtime" to run the feature processor. The other tests will work even if this argument is not set.

Writing the tests is actually very simple, there is a "JUnit 4 in 60 seconds" tutorial here:
http://www.cavdar.net/2008/07/21/junit-4-in-60-seconds/

3. Status

The code runs OK (tested with the old non-HTTP server+client) with a modified version of the voice hsmm-slt, which I will send to the group by email shortly. The code has not been tested with any unit selection voice yet.

3. Minor optimisations

I have run a profiler to see where the time is spent at startup; it turns out the two main sources are:

- loading the FreeTTS English lexicon (we will replace that with our FST-based lexicon, so no worry);
- loading the HTS tree files. Here, I have managed to get a 30% increase in speed by replacing the use of Scanner (which does expensive regular expression matching) with BufferedReader.readLine() and StringTokenizer. 


git-svn-id: https://mary.opendfki.de/repos/trunk@1511 953a6561-930b-0410-b2a6-db37d1b2ae63