Cleanup: Merged the two phone set implementations into one. This fixes ticket #177.
IMPORTANT: In order to run the system now, you need to use the new english.config and you need to make sure you have lib/modules/en/us/lexicon/allophones.en_US.xml in your runtime version of MARY.


Details
-------

Before, we used marytts.features.PhoneSet in the feature processor code, and marytts.modules.phonemiser.PhonemeSet in the rest of the code.

This changeset replaces both with marytts.modules.phonemiser.AllophoneSet, which provides both APIs, and also a getSilence() method to determine the silence symbol used in the given Allophone set.

AllophoneSet reads from an XML file in a new format, defining each allophone as an XML element, such as:
	<vowel ph="A" vlng="l" vheight="3" vfront="3" vrnd="-"/>
	<consonant ph="p" ctype="s" cplace="l" cvox="-"/>

The list of features to annotate on the allophones is given in the top level element's "features" attribute:

<allophones name="sampa" xml:lang="en-US"
		features="vlng vheight vfront vrnd ctype cplace cvox">

This makes it possible to use a different set of features for a language where that makes sense. For example, for Tibetan, we can replace "cvox" (voicedness of consonants, which is not distinctive in Tibetan) with "casp" (aspiration of consonants).

The new allophone files are at the following places:
US English: lib/modules/en/us/lexicon/allophones.en_US.xml
Tibetan: lib/modules/tib/lexicon/allophones.tib.xml
German: lib/modules/de/lexicon/allophones.de.xml

The old XML file, which was underlying PhonemeSet, is still used for the class KlattDurationModeller.KlattDurationParams, which provides minimum and inherent durations for all allophones.

git-svn-id: https://mary.opendfki.de/repos/trunk@1517 953a6561-930b-0410-b2a6-db37d1b2ae63