Defined a new, binary file format for CART files, and documented it in doc/UnitDatabaseFormat.odt.
The new file uses truly binary values throughout, instead of text representations; as a result, it is now about 1/4 of the size it had before.

Also, the binary CART file now includes the feature definition, so that a CART file is now a self-contained file.

The text format produced by MaryCARTWRiter.toTextOut() is unchanged compared to before. However, MaryCARTReader can only read the binary file format, i.e. the text format is for human inspection only.

This has been tested only with StringAndFloatLeafNodes for the moment, using CMUDict2MaryFST.

git-svn-id: https://mary.opendfki.de/repos/trunk@1534 953a6561-930b-0410-b2a6-db37d1b2ae63