Change to the http server code, to simplify the structure of the interaction with a web server and to make the information requests more intuitive.
Uses AJAX to request information from the server and put it into the right places without ever reloading the entire page. Audio is embedded via the object tag containing a direct link to the synthesis request. No multi-pass generation of audio data needed anymore.

Changed information requests -- see MaryHttpServer javadoc for a short description.

This is tested with Firefox, Safari, Internet Explorer and seems to work more or less OK.

NOTE: The MaryHttpClient is broken in this version! Will be fixed soon.

git-svn-id: https://mary.opendfki.de/repos/trunk@1664 953a6561-930b-0410-b2a6-db37d1b2ae63