Enhancement, ticket #190: The client can now request feature vectors containing a specifiable list of features.
First of all, the client can request a list of available features for a given voice or locale, e.g.
	http://localhost:59125/features?locale=en

Then, I introduced the concept of "output type parameters", which the client can optionally send along with the process request. These output type parameters, if present, will be set in the MaryData object sent as input to the MaryModule producing the output type. In the case of TARGETFEATURES, that is the TargetFeatureLister, and it interprets the parameters as a list of features to produce.


To try, use the command line interface of MaryHttpClient:

echo "Hallo" |java -Dserver.host=localhost -Dinput.type=TEXT -Dtput.type=TARGETFEATURES "-Doutput.type.params=phoneme stressed unit_duration" marytts.client.http.MaryHttpClient 

git-svn-id: https://mary.opendfki.de/repos/trunk@1678 953a6561-930b-0410-b2a6-db37d1b2ae63