Implemented simple polynomial fitting using a least squares method.
Added support for additional data series to FunctionGraph. The X range and number of points must be
the same as for the primary data series; y range is adapted. Presentation style of the primary data series can be adapted using setPrimaryDataSeriesStyle(); for additional series, style is set
when calling addDataSeries().

Should maintain compatibility with existing code.

git-svn-id: https://mary.opendfki.de/repos/trunk@1775 953a6561-930b-0410-b2a6-db37d1b2ae63