Substantial cleanup, fixing ticket #208: Extracted abstract superclass MaryClient, with two subclasses MarySocketClient and MaryHttpClient.
Ease of use: you can use the static method
MaryClient.getMaryClient()
to create either a MarySocketClient or a MaryHttpClient, simply by trying. In other words, the client code can treat the MaryClient as the same kind of entity no matter what the communication protocol between client and server.

Fixed a number of issues with effects on the socket protocol on the way. 

There is now only one GUI for the client, the MaryGUIClient, which uses MaryClient.getMaryClient() and so works with both protocols.

git-svn-id: https://mary.opendfki.de/repos/trunk@1897 953a6561-930b-0410-b2a6-db37d1b2ae63