Several important fixes to REALISED_ACOUSTPARAMS and REALISED_DURATIONS code, triggered by ticket:169:
- REALISED_ACOUSTPARAMS now follows the same conventions as ACOUSTPARAMS, i.e. phone durations are given in integer milliseconds, whereas end times are given in float seconds.

THIS IS AN INCOMPATIBLE CHANGE! End times were output for unit selection voices in milliseconds in REALISED_ACOUSTPARAMS before, even though they were in seconds in ACOUSTPARAMS. Any code using the REALISED_ACOUSTPARAMS output format will have to modify interpretation, but it still seems like the right thing to do. In particular, this applies to the SpeechSynthesis components in the SEMAINE system.

- REALISED_DURATIONS now uses the end attributes, which are more accurate than the integer millisecond duration attributes. Over a normal utterance, rounding errors did formerly sum up to several milliseconds.

- For HMM-based voices, REALISED_ACOUSTPARAMS was not working properly: end times were always 0, and boundary durations were ignored. Both is fixed now, so REALISED_ACOUSTPARAMS should be usable for all types of voices now.

git-svn-id: https://mary.opendfki.de/repos/trunk@1962 953a6561-930b-0410-b2a6-db37d1b2ae63