fixed bug in database.config handling
In database.config, a path with no trailing slash is adjusted correctly, but this bug inserted an additional, malformed key, which could trigger a StringIndexOutOfBoundsException.

Probable origin of the bug is copy/pasting from one of several similar conditional blocks in DatabaseLayout.java, which use subtly different variable names for the same semantics. Perhaps they should be merged into a single function.

git-svn-id: https://mary.opendfki.de/repos/trunk@2273 953a6561-930b-0410-b2a6-db37d1b2ae63