patched handling of externally installed EST
ESTCaller ignores a proper installation of Edinburgh Speech Tools (EST), as well as the canonical ESTDIR environment variable, relying solely on paths in MARY's config files.

While the path of a local EST build can be provided to MARY, an *installation* of EST (e.g. the speech-tools package in Debian/Ubuntu repositories) places EST binaries in /usr/bin. Voice Import Tools can be persuaded to use those installed binaries by setting *.estDir = /usr/ in the configs, but ESTCaller is hardcoded to look in $ESTDIR/main, which would throw several exceptions. This behavior is now modified to look in $ESTDIR/bin instead. This patch will continue to work for local builds of EST, since the $ESTDIR/bin directory contains symlinks to the binaries in $ESTDIR/main.

git-svn-id: https://mary.opendfki.de/repos/trunk@2275 953a6561-930b-0410-b2a6-db37d1b2ae63