re-enabled Festvox2MaryTranscripts and SnackPitchmarker components in voice import
Festvox2MaryTranscripts may be required for people using arctic data. It is documented in the wiki, but unavailable in the Voice Import Tools, which may present an obstacle to users who stick to using the GUI. Enabled again.

SnackPitchmarker is possibly less critical, because an alternative is available (PraatPitchmarker). However, it is also documented, so should perhaps be available to users who prefer to use the ESPS algorithm.

git-svn-id: https://mary.opendfki.de/repos/trunk@2279 953a6561-930b-0410-b2a6-db37d1b2ae63