centralized configuration of pitchmark file extension and directory
Replaced all component-specific pitchmark file references (".pm" and ".pm.corrected" file extensions, "pm" directory, local properties, and especially hardcoded variables) with two new global properties, "db.pmDir" and "db.pmExtension".

This fixes a bug in BasenameTimelineMaker, which caused the uncorrected pitchmarks to be used.

This also removes the distinction between the concepts of "corrected" and "uncorrected" pitchmarks. The rationale is that any pitchmark correction should be performed by the pitchmarker components, and this should be opaque to later components, which should only use the correct pitchmarks anyway.


git-svn-id: https://mary.opendfki.de/repos/trunk@2286 953a6561-930b-0410-b2a6-db37d1b2ae63