patched concurrent modification bug in basename iteration
The for loop using an index variable to iterate over the BasenameList triggered a classic concurrent modification problem. It didn't throw an IndexOutOfBoundsException because the break condition calls bnl.getLength() with each iteration. *However*, the iterator i was not decremented after a remove(), so the following basename was skipped when it should have been processed; the result was that in a set of missing lab files, half of them were not removed, causing problems down the line.

Patched the bug by decrementing the iterator after the remove(), but moved i from the compute() method to the class variable basenameIndex in order to expose it to computePhoneLabel(). Very dirty.

IMHO, this iterating and basename removing should really be done using an Iterable BasenameList, and the iterator's built-in remove() method. In fact, the whole basename checking and removing shouldn't be done as a side-effect of compute() at all, but in a proper file-testing method.

git-svn-id: https://mary.opendfki.de/repos/trunk@2289 953a6561-930b-0410-b2a6-db37d1b2ae63