added style-specific basename-handling to AllophonesExtractor
This is done by setting up a style definition file (disabled by default), which lists basenames or glob expressions and corresponding styles. Each basename is matched against this definition and the corresponding style applied to the generated allophones. This change introduces a new file reader for this kind of glob-based basename classification, and reuses the glob matching code from Apache Jakarta ORO.

git-svn-id: https://mary.opendfki.de/repos/trunk@2303 953a6561-930b-0410-b2a6-db37d1b2ae63