better blacklisting
we no longer get a 90-fold performance drop during active blacklisting. In addition, it is now possible to specify a blacklist attribute containing one or more basename strings *anywhere* in the MaryXML, so that units from given basenames can be excluded locally, with arbitrary scope.

WARNING: this whole blacklisting hack is experimental and will be removed in the future, to be replaced by a <debug> tag, after which the same results can be achieved with valid MaryXML.

git-svn-id: https://mary.opendfki.de/repos/trunk@2329 953a6561-930b-0410-b2a6-db37d1b2ae63