Substantially revised the internals of TimelineReader. Changed from reading with a RandomAccessFile to using a FileChannel/MappedByteBuffer.
Reading does not to be synchronized anymore; in addition, the memory map access is simply much faster.

As a result, we get about 4 times faster unit selection than before.

Tested this with cmu-slt voice and DemandingClient, seems ok. JUnit Timeline test also running, and a quick test with BasenameTimelineWriter also runs through without complaining.

Remains to be seen how this reacts with a lot of voices.

git-svn-id: https://mary.opendfki.de/repos/trunk@2382 953a6561-930b-0410-b2a6-db37d1b2ae63