AcousticModeller can now replace DummyAllophones2AcoustParams
Voice constructor now checks voice config for acoustic model definitions and provides them in a Map, which is queried by AcousticModeller.

AcousticModeller should now be suited to replace DummyAllophones2AcoustParams in the modules.classes.list of marybase.config. However, in that case, no voices are loaded yet, so AcousticModeller has a delayedStartup() method which is triggered the first time a Request calls for a voice without conventional preferredModules. If all goes well, this method pulls in the actual Models and completes the module initialization. Otherwise, it passes out unmodified MaryData.

TODO:
* Make feature processors access the same getters to have flexible GenericContinuousFeatureProcessors


git-svn-id: https://mary.opendfki.de/repos/branches/vq_unitselection@2395 953a6561-930b-0410-b2a6-db37d1b2ae63