more polishing of acoustic model loading
Models now have their TargetFeatureComputers set by a new FeatureProcessorManager(Voice) constructor.

Note that this changeset breaks voices built with previously hardcoded vq_unitselection-specific voice quality feature processors. Could've @Deprecated them, but we won't want that code in the trunk anyway, so it had to break sooner or later.

Currently the Locale determines whether the vanilla FPM is contructed, or the one from language.de.features, which calls deprecated methods.

TODO: If the voice config does not define acoustic models for all continuous features in the halfphone feature file's FeatureDefinition, an IllegalArgumentException is thrown by source:/branches/vq_unitselection/java/marytts/features/TargetFeatureComputer.java@2332:68#L47
This is probably undesirable, because a voice config that is edited or mangled so that it fails to define all acoustic models will break a MARY installation, and the only way to figure out what's missing from the voice config and fix it is to actually inspect the feature file's FeatureDefinition somehow. We may get around this by relaxing the Exception handling if that feature has zero weight, but not in this changeset...


git-svn-id: https://mary.opendfki.de/repos/branches/vq_unitselection@2408 953a6561-930b-0410-b2a6-db37d1b2ae63