FD-PSOLA unit concatenator now adjusts durations in reference to phone segments
Predicted target durations are now (almost) achieved, and the following pitfall avoided:

The unit duration feature processor predicts halfphone unit durations as exactly half of the corresponding phone duration.

However, the two realized duration of the halves of a phone as built from diphones will almost never have a one-to-one ratio; often it will be extremely skewed. Applying FD-PSOLA resynthesis to the halfphone units in such a case will either produce serious artifacts, or one can clip the scaling factors to a "safe" range, but thereby fail to achieve the target duration.

By calculating the scaling factor from the full phone, not each halfphone individually, the halfphone units are modified relative to their proportion of the phone duration, and the scaling factors will be mostly in the safe range anyway (assuming reasonable coverage in the voice data).


git-svn-id: https://mary.opendfki.de/repos/branches/vq_unitselection@2415 953a6561-930b-0410-b2a6-db37d1b2ae63