fix for ticket:275
FD-PSOLA processing now updates the durations of Datagrams. These updated durations are propagated to the SelectedUnits' ConcatenationData, which automagically updates the MaryXML in [source:/trunk/java/marytts/unitselection/UnitSelectionSynthesizer.java@2420:225,257-258#L219 UnitSelectionSynthesizer].

The drawback is that such a processed Datagram's duration does not respect its data, but this should not cause problems, since Datagrams do not seem to be used after this point, only the raw audio samples.

git-svn-id: https://mary.opendfki.de/repos/trunk@2447 953a6561-930b-0410-b2a6-db37d1b2ae63