quick fix for model file overwriting in HMM voice building on case insensitive filesystems
Filesystems like NTFS and HFS+ with default parameters prevent files with names like `a` and `A` from co-existing. This means that in `VOICEDIR/hts/models/qst001/ver1/cmp/HRest/`, the file for model `a` will be overwritten by the file for model `A` (or vice-versa).

This fix extends the code already in place to map "tricky" allophone names (which can lead to problems when using them as filenames) to safe strings, so that every non-lower-case allophone name will also be remapped if it differs only in case from that of another allophone in the AllophoneSet whose name is lower-case.

Also disabling the currently unsupported HMMVoiceMakeVoiceAdapt component (and removing a random line of debug code).


git-svn-id: https://mary.opendfki.de/repos/trunk@2472 953a6561-930b-0410-b2a6-db37d1b2ae63