cleanup and restructuring in AcousticModeller and modules.acoustic
Models now have `predictFrom` and `applyTo` properties, which default to "segments" unless set differently in the voice config.

Unit-selection voices use "F0" for `leftF0`; `midF0` and `rightF0` are put in "otherModels" (this workaround seems like a good compromise, since the most generic case should use only one model to predict f0). This means it is possible that rightF0 could be applied after midF0 (because the order of keys in otherModels is random), but (if it occurs) that should not affect performance (otherwise that would be a bug which we can fix in one of several ways).

HMM-specific stuff should only appear in HMMModel; once that is achieved, we should replace the `applyTo` and `applyFromTo` methods in Model with a simple `apply`, which knows from the `predictFrom` and `applyTo` properties which `elementLists` to process.

git-svn-id: https://mary.opendfki.de/repos/trunk@2496 953a6561-930b-0410-b2a6-db37d1b2ae63