fix for ticket:292
HnmVoiceInstaller was 99% duplicate code from (old version of) VoiceInstaller; added two properties to VoiceInstaller and removed HnmVoiceInstaller.

To install an HNM voice, now simply use the regular VoiceInstaller, but set `CONCATENATORCLASS` to `HnmUnitConcatenator` and `TIMELINEREADERCLASS` to `HnmTimelineReader`, and change `WAVETIMELINE` to point to the HNM timeline created with the HnmTimelineMaker component (which however is still disabled by default).

Incidentally, the user can now also configure a unit-selection voice to use FD-PSOLA modification by setting `CONCATENATORCLASS` to `FdpsolaUnitConcatenator`.

This change should be documented in the Wiki.

git-svn-id: https://mary.opendfki.de/repos/trunk@2519 953a6561-930b-0410-b2a6-db37d1b2ae63