Small task, big effort needed to resolve it: ticket:305 represents a NullPointerException when sending a RAWMARYXML input for a British English voice for which no voice exists (as when building the first British English voices).
The solution involves getting the AllophoneSet for the locale of the input data. To make things more complicated, this is not the locale of the processing components.

Used this opportunity to clean up suboptimal Exception handling in the context of AllophoneSet: it now throws a MaryConfigurationException if something that *should* be available cannot be loaded properly.

git-svn-id: https://mary.opendfki.de/repos/trunk@2605 953a6561-930b-0410-b2a6-db37d1b2ae63