unmerged r2213
EHMMLabeler is not only used for HMM voices, and should therefore not necessarily depend on `MARYBASE/lib/external/externalBinaries.config`. In any case, the user should have the possibility of specifying the location of `ehmm` in the GUI. This change restores configurability of `EHMMDIR`, but queries `externalBinaries.config` for the default path if none is provided and `externalBinaries.config` exists.

git-svn-id: https://mary.opendfki.de/repos/trunk@2621 953a6561-930b-0410-b2a6-db37d1b2ae63