improved Exception handling in AcousticModeller and parts of modules.acoustic
Also, I've removed the `type` field from the Models; it was used only to determine which class to instantiate, and by the time we know this (in [source:/trunk/java/marytts/modules/synthesis/Voice.java@2651:264-281#L254 Voice#loadAcousticModels]), it's served its purpose and does not need to be dragged along into the Model itself.

In addition, I've modified the ProsodyElementHandler a tiny bit to allow other classes to call getF0Contour().


git-svn-id: https://mary.opendfki.de/repos/trunk@2690 953a6561-930b-0410-b2a6-db37d1b2ae63