restructuring of FD-PSOLA unit concatenator
Most significantly, we now have a ProsodyAnalyzer, which records the predicted and realized prosodic parameters returned from the unit selection process, and provides high-level, phone-based methods which can be used to generate the modification parameters, or just analysis.

Note that some refactoring will take place when this is reshuffled to mesh better with the rest of the code.


git-svn-id: https://mary.opendfki.de/repos/trunk@2691 953a6561-930b-0410-b2a6-db37d1b2ae63