Attempt to apply our new error handling principles to the startup code of TimelineReader. I have gone for the fundamental guarantee only here.
Since I have applied the error handling principles also to the classes that are called in TimelineReader.load(), things did spread rather widely.

An interesting case is MaryHeader. Applying the fundamental guarantee here meant that the tests for a valid Mary header had to be called in the constructor directly. That meant removing the public method isMaryHeader(), which had been called from many places, usually after creating the MaryHeader object. Now we can rely (and the constructor's javadoc makes that explicit) that, if a MaryHeader object is created, it is guaranteed to be a valid mary header, i.e. to have a legal magic number and a legal type.

Another interesting case was TranscriptionGUI. Here I added code to inform the user of an error condition when an exception was thrown in the business logic code, with a popup alerting the user and full detail on the console.

A third interesting issue was the question whether the constructors of the local classes TimelineReader.Index and TimelineReader.ProcHeader should be treated according to the "internal business logic" or the "interface between components" rules. The local classes and their constructors were public, which indicated "interface" rules; but that felt wrong, since they were created only from within the TimelineReader. The solution I adopted was to leave the classes as public (needed because they are used here and there) but to make their constructors private, allowing me to treat them according to "internal business logic" rules.

Still need to adapt most of the methods in TimelineReader to the new error handling; hopefully that will have fewer repercussions throughout the code.

git-svn-id: https://mary.opendfki.de/repos/trunk@2698 953a6561-930b-0410-b2a6-db37d1b2ae63