added energy-based transient halfphone splitting
If `HalfPhoneUnitLabelComputer.energyBasedTransientSplitting` is set to `true`, an energy analysis file will be created for each wav file, and the midpoint (i.e. halfphone boundary) of each transient phone (i.e. plosives) is determined by finding the frame with the highest increase in RMS energy within the phone boundaries. Results seem encouraging for bits3 data, but mixed for pavoque data (may have to do with labeling), so I'm leaving the default at `false`.


git-svn-id: https://mary.opendfki.de/repos/trunk@2702 953a6561-930b-0410-b2a6-db37d1b2ae63