possible fix for ticket:357
Now we properly handle a HeadlessException thrown by the pointlessly gooey dependency error message in a headless environment. Not sure why it should be graphical in the first place.

We seem to have to resort to STDERR, because the logging subsystem is not yet initialized. (I'm not sure why MaryProperties.readProperties() needs to be called twice in Mary.java...)

git-svn-id: https://mary.opendfki.de/repos/trunk@2718 953a6561-930b-0410-b2a6-db37d1b2ae63