fix for ticket:340
We now include `mary-mbrola.jar` with the `marybase` pack, so that we don't lose control over the MBROLA wrapper code to voices which install a buggy version of that JAR. Simultaneously, the InstallerGUI is modified so that it does not overwrite existing JAR files if they are newer (this is determined by comparing the "Specification-Version" in the JAR manifests).

Tested on Snow Leopard, Windows 2003, Ubuntu Maverick.

Note that this fix still leaves a few rough edges:
- It's possible to craft a voice or language component that will break a Mary installation by overwriting JARs or removing them at uninstall. This is nothing new, but it's not solved here, either.
- Generally it's a bad idea to share files between more than one component, but there's still no proper guard against such practice in place.


git-svn-id: https://mary.opendfki.de/repos/trunk@2783 953a6561-930b-0410-b2a6-db37d1b2ae63