better solution for ticket:379
Modification is now passed in as OUTPUT_TYPE_PARAMS=MODIFICATION. Since this clashed with the previous form element which was used only for target feature listing, the way both are handled has been updated. Note that this is rather kludgy and may cause annoyance if OUTPUT_TYPE_PARAMS is extended for other uses in the future.

Output type parameters are now passed through all the way to the Synthesizers (cf. r2816). This should not have any side effects.

The Synthesis module is essentially back at r2541, except for the parameter passing.

The UnitSelectionVoice still has an additional UnitConcatenator, but only once the first MODIFICATION is requested.

The UnitSelectionSynthesizer checks for MODIFICATION in the output type parameters, and uses the appropriate UnitConcatenator from the Voice.


git-svn-id: https://mary.opendfki.de/repos/branches/4.3-branch@2817 953a6561-930b-0410-b2a6-db37d1b2ae63