`boundary` elements no longer receive an (incorrect) "end" attribute after rate modification
For consistency with the inconsistent attribute naming already in place, boundaries now use the "duration" attribute, and do not get an "end" attribute. Incidentally, the "end" attribute was previously computed under the false assumption that the durations of boundaries are modified if "rate" is specified in the enclosing `prosody` element (however, see ticket:387), which caused the insertion of incorrect values.

git-svn-id: https://mary.opendfki.de/repos/trunk@2879 953a6561-930b-0410-b2a6-db37d1b2ae63