improvements to Tcl client for MaryTTS
This is a ptch to the tcl client I wrote.  It adds the -underline 0 to
the menus which ought to give the <alt-F> to access the file menu, etc,
(only I am finding on Vista that it does not) and it also adds the
ability to have much of the GUI, apart from the Tk_messageBox for the
help, in large print.

Hugh