check for infinity (=19) faster than before
The problem was that the numbers are calculated precisely and only
afterwards is the returned value limited to a maximum of 19 (for
compatibility with flite). It's not really worth finding out that
the utterance contains 92 words when we then `rail' to 19 anyways,
especially given that DOM traversal is relatively slow.