Bugfix german AbbrevEP
Integers were expanded without an MTU around them, making it
impossible to reconstruct the original text.

The expanded integer is now wrapped in an MTU.  This also applies to
decimal numbers.

Testcases were updated to reflect the change in the *.preprocessed
files.  A new testcase for the described problem was added.